+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Format: [Keep a Changelog](https://keepachangelog.com/)
 
 ## [Unreleased]
 
+### Added
+- Show a non-layout-affecting progress bar at the top of the main panel during unified data refreshes
+
 ## [0.10.2] - 2026-08-04
 
 ### Added